refac: Move mock_agent_module function to the integration conftest #2537
