+++ tests/integration/test_vulnerability_detector/conftest.py
@@ -66,18 +66,6 @@ def mock_agent():
 
 
 @pytest.fixture(scope='module')
-def mock_agent_module():
-    """
-    Fixture to create a mocked agent in wazuh databases
-    """
-    agent_id = create_mocked_agent(name="mocked_agent")
-
-    yield agent_id
-
-    delete_mocked_agent(agent_id)
-
-
-@pytest.fixture(scope='module')
 def check_cve_db():
     """
     Fixture to check if the CVE database exists and its tables are created
